+++ src/test/java/com/codefork/refine/orcid/OrcidSmartNamesTest.java
@@ -17,6 +17,7 @@ import org.springframework.test.web.servlet.MvcResult;
 
 import java.util.Properties;
 
+import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.mockito.Mockito.mock;
@@ -24,8 +25,8 @@ import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 
 @RunWith(SpringRunner.class)
-@WebMvcTest(Orcid.class)
-public class OrcidTest {
+@WebMvcTest(OrcidSmartNames.class)
+public class OrcidSmartNamesTest {
 
     @TestConfiguration
     static class TestConfig {
@@ -42,7 +43,7 @@ public class OrcidTest {
             props.put("name", "ORCID");
 
             Config config = mock(Config.class);
-            when(config.getDataSourceProperties("orcid")).thenReturn(props);
+            when(config.getDataSourceProperties("orcidsmartnames")).thenReturn(props);
             return config;
         }
 
@@ -53,28 +54,28 @@ public class OrcidTest {
     }
 
     @Autowired
-    Orcid orcid;
+    OrcidSmartNames orcid;
 
     @Autowired
     MockMvc mvc;
 
     @Test
-    public void testOrcidServiceMetaData() throws Exception {
-        MvcResult result = mvc.perform(get("/reconcile/orcid")).andReturn();
+    public void testSmartNamesServiceMetaData() throws Exception {
+        MvcResult result = mvc.perform(get("/reconcile/orcid/smartnames")).andReturn();
 
         String body = result.getResponse().getContentAsString();
 
         JsonNode root = new ObjectMapper().readTree(body);
 
-        assertEquals("ORCID", root.get("name").asText());
+        assertEquals("ORCID - Smart Names Mode", root.get("name").asText());
     }
-
+    // https://github.com/codeforkjeff/conciliator/issues/8
     @Test
-    public void testLiveSearch() throws Exception {
+    public void testLiveSearchSmartNames() throws Exception {
 
-        String json = "{\"q0\":{\"query\": \"stephen hawking\",\"type\":\"/people/person\",\"type_strict\":\"should\"}}";
+        String json = "{\"q0\":{\"query\": \"Igor Ozerov\",\"type\":\"/people/person\",\"type_strict\":\"should\"}}";
 
-        MvcResult mvcResult = mvc.perform(get("/reconcile/orcid").param("queries", json)).andReturn();
+        MvcResult mvcResult = mvc.perform(get("/reconcile/orcid/smartnames").param("queries", json)).andReturn();
 
         String body = mvcResult.getResponse().getContentAsString();
 
@@ -82,13 +83,22 @@ public class OrcidTest {
 
         JsonNode results = root.get("q0").get("result");
 
-        assertEquals(3, results.size());
+        assertEquals(1, results.size());
 
         JsonNode result1 = results.get(0);
-        assertEquals("Stephen Hawking", result1.get("name").asText());
+        assertEquals("Igor OZEROV", result1.get("name").asText());
         assertEquals("Person", result1.get("type").get(0).get("name").asText());
-        assertEquals("0000-0002-9079-593X", result1.get("id").asText());
+        assertEquals("0000-0001-5839-7854", result1.get("id").asText());
         assertFalse(result1.get("match").asBoolean());
     }
 
+    @Test
+    public void testParseName() {
+        assertArrayEquals(OrcidSmartNames.parseName("joe schmoe"),
+                new String[] { "joe", "schmoe" });
+        assertArrayEquals(OrcidSmartNames.parseName("schmoe, joe"),
+                new String[] { "joe", "schmoe" });
+        assertArrayEquals(OrcidSmartNames.parseName("dr. joe schmoe"),
+                null);
+    }
 }
